+++ clients/Android/app/src/test/java/org/mtuosc/techchat/asynctasks/SignUpUserTaskTest.java
@@ -1,0 +1,11 @@
+package org.mtuosc.techchat.asynctasks;
+
+import static org.junit.Assert.*;
+
+/**
+ * Created by ryan on 2/4/18.
+ */
+public class SignUpUserTaskTest {
+
+
+}
